+++ CHANGELOG.md
@@ -5,13 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.4.3] - 2026-04-03
+
+### Added
+
+- `headers` property to `DownloadResult` with the `Last-Modified` HTTP response
+  header from filter downloads [AdguardBrowserExtension#3407].
 
 ### Changed
 
 - Updated [@adguard/diff-builder] to 1.1.4.
 
-[Unreleased]: https://github.com/AdguardTeam/FiltersDownloader/compare/v2.4.2...HEAD
+[2.4.3]: https://github.com/AdguardTeam/FiltersDownloader/compare/v2.4.2...v2.4.3
+[AdguardBrowserExtension#3407]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3407
 
 ## [2.4.2] - 2025-06-30